edit

+++ lib/tasks/remove_image.rake
@@ -1,13 +1,14 @@
 namespace :image do
   task :remove_image, [:tz] => :environment do |t, args|
-    system("rm -rf public/BEAU")
-    system("rm -rf public/CNP")
-    system("rm -rf public/CLAB")
-    system("rm -rf public/LABO")
-    system("rm -rf public/MART")
-    system("rm -rf public/TMR")
-
-    notifier = Slack::Notifier.new "https://hooks.slack.com/services/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"
-    notifier.ping "CNP Image Remove Execute !!!"
+    if args[:tz] == "Seoul"
+      system("rm -rf public/BEAU")
+      system("rm -rf public/CNP")
+      system("rm -rf public/CLAB")
+      system("rm -rf public/LABO")
+      system("rm -rf public/MART")
+      system("rm -rf public/TMR")
+      notifier = Slack::Notifier.new "https://hooks.slack.com/services/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"
+      notifier.ping "CNP Image Remove Execute !!!"
+    end
   end
 end
